Fixed remaining building issues when building to UWP EXCEPT for the ObjectCreator.cs

+++ lang/Unity3D/Assets/avro-master/src/apache/main/Generic/GenericWriter.cs
@@ -278,7 +278,7 @@ namespace Avro.Generic
         /// <returns>The array element at the index</returns>
         protected virtual object GetArrayElement(object value, long index)
         {
-            return (value as Array).GetValue(index);
+            return (value as Array).GetValue((int)index);
         }
 
         /// <summary>
